refactor: 컴포넌트 분리 구조 변경
