+++ src/components/SearchBar/SearchBar.jsx
@@ -3,10 +3,10 @@ import searchIcon from 'assets/icons/search.svg';
 
 function SearchBar() {
   return (
-    <S.SearchBarContainer>
+    <S.Container>
       <S.Input placeholder='링크를 검색해보세요.' />
       <S.Icon src={searchIcon} alt='검색 아이콘' />
-    </S.SearchBarContainer>
+    </S.Container>
   );
 }
