+++ src/pages/Folder/Folder.jsx
@@ -1,14 +1,45 @@
+import { useSearchParams } from 'react-router-dom';
+import * as S from './Folder.style';
 import Layout from 'components/Layout';
-import AddLink from './components/AddLink';
-import LinkList from './components/LinkList';
+import useAsync from 'hooks/useAsync';
+import { getLinks } from 'utils/apiClient';
+import AddLinkContainer from './components/AddLinkContainer';
+import SearchBar from 'components/SearchBar';
+import FoldersContainer from './components/FoldersContainer';
+import CardsContainer from 'components/CardsContainer';
+import NoLinkView from './components/NoLinkView';
 
 const DEFAULT_USER_ID = 1;
 
 function Folder() {
+  const [linksData, , , getLinksAsync] = useAsync({
+    asyncFunction: getLinks,
+    initialArgs: DEFAULT_USER_ID,
+  });
+
+  const [searchParams, setSearchParams] = useSearchParams();
+
+  const setFolderLinks = (folderId) => {
+    setSearchParams({ folderId } ?? {});
+    getLinksAsync(DEFAULT_USER_ID, folderId);
+  };
+
   return (
     <Layout isLoggedIn userId={DEFAULT_USER_ID}>
-      <AddLink />
-      <LinkList userId={DEFAULT_USER_ID} />
+      <AddLinkContainer />
+      <S.ContentContainer>
+        <SearchBar />
+        <FoldersContainer
+          userId={DEFAULT_USER_ID}
+          setFolderLinks={setFolderLinks}
+        />
+
+        {linksData?.data.length !== 0 ? (
+          <CardsContainer cards={linksData?.data} />
+        ) : (
+          <NoLinkView />
+        )}
+      </S.ContentContainer>
     </Layout>
   );
 }
