+++ src/pages/Folder/Folder.style.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import { onPc } from 'styles/mediaQuery';
+
+export const ContentContainer = styled.main`
+  padding: 2rem 3.2rem;
+
+  ${onPc} {
+    margin: 0 auto;
+    width: 115rem;
+  }
+`;
