--- src/pages/Folder/components/AddLink/index.js
@@ -1,1 +1,0 @@
-export { default } from './AddLink';
