+++ src/pages/Folder/components/AddLinkContainer/index.js
@@ -1,0 +1,1 @@
+export { default } from './AddLinkContainer';
