--- src/pages/Folder/components/FolderHeader/index.js
@@ -1,1 +1,0 @@
-export { default } from './FolderHeader';
