+++ src/pages/Folder/components/FoldersContainer/index.js
@@ -1,0 +1,1 @@
+export { default } from './FoldersContainer';
