+++ src/pages/Folder/components/FoldersView/index.js
@@ -1,0 +1,1 @@
+export { default } from './FoldersView';
