--- src/pages/Folder/components/LinkList/LinkList.style.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-import { onPc } from 'styles/mediaQuery';
-
-export const ContentContainer = styled.main`
-  padding: 2rem 3.2rem;
-
-  ${onPc} {
-    margin: 0 auto;
-    width: 115rem;
-  }
-`;
