--- src/pages/Folder/components/LinkList/index.js
@@ -1,1 +1,0 @@
-export { default } from './LinkList';
