--- src/pages/Folder/components/NoLink/NoLink.js
@@ -1,7 +1,0 @@
-import * as S from './NoLink.style';
-
-function NoLink() {
-  return <S.NoLink>저장된 링크가 없습니다.</S.NoLink>;
-}
-
-export default NoLink;
