--- src/pages/Folder/components/NoLink/index.js
@@ -1,1 +1,0 @@
-export { default } from './NoLink';
