+++ src/pages/Folder/components/NoLinkView/NoLinkView.js
@@ -1,0 +1,7 @@
+import * as S from './NoLinkView.style';
+
+function NoLinkView() {
+  return <S.NoLink>저장된 링크가 없습니다.</S.NoLink>;
+}
+
+export default NoLinkView;
