+++ src/pages/Folder/components/NoLinkView/NoLinkView.style.js
