+++ src/pages/Folder/components/NoLinkView/index.js
@@ -1,0 +1,1 @@
+export { default } from './NoLinkView';
